Set completed workflow current version to lastWriteVersion (#4431)

+++ service/history/execution/context.go
@@ -953,7 +953,7 @@ func (c *contextImpl) PersistStartWorkflowBatchEvents(
 			// TransactionID is set by shard context
 		},
 	)
-	return int64(size), err
+	return size, err
 }
 
 func (c *contextImpl) PersistNonStartWorkflowBatchEvents(
